requirements.txt
mysql-connector-python
psycopg2-binary
pymysql
cryptography